error handling

+++ pipelines/original/collect_records.py
@@ -80,9 +80,11 @@ with open("failed_records.txt") as file:
             
             if os.path.exists(records_path):
                 shutil.move(records_path, f"failed_records/sims_data_records/{map_id}")
-                
-            with open(f"failed_records/dnn_input_data/{map_id}/{sim_id}/params.yaml", "a") as file:
-                yaml.dump(f"failed_reason: {failed_reason}", file)
+            
+            params_path = f"failed_records/dnn_input_data/{map_id}/{sim_id}/params.yaml"
+            if os.path.exists(params_path):
+                with open(params_path, "a") as file:
+                    yaml.dump(f"failed_reason: {failed_reason}", file)
 
 with open("failed_records.txt", 'w') as f:
     f.write('fail_reason,map_name,sim_id\n')
